Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

ICS kernel source (4.1.B.0.431) broke the touch screen drivers on Iyokan making it only update every 10 pixel. 
I have copied the driver from the GB source and enabled gesture in touch-iyokan (needed for it to work)
which now brings the touch screen back to working state again. The difference is like day and night.

Instead of modifying the current driver, risking it to break on other devices, I just created another config option (+ a few dependency configs that was needed)

Hopefully Sony has solved the issue in the new ICS update (when we get the source) but until then our Iyokan users can
finally use the touch screen without the choppines  again.

If patch is accepted I will submit it to the CM9 branch too since it's broken there as well after the update to 4.1.B.0.431 source.

http://code.google.com/p/freexperia/issues/detail?id=766 and http://code.google.com/p/freexperia/issues/detail?id=493
describes the problem in more detail.

I have been tested it for a few hours on my mk16i and it works great.

Patch-set: 1