Update patch set 1
Patch Set 1:

ICS kernel source (4.1.B.0.431) broke the touch screen drivers on Iyokan making the touch precision set to 9 pixels.

It turns out that the screen needs the gesture setup during init to be able to report every update from the touch screen to the driver. The precision can be set with act_dist. Leaving it set to 1 pixels makes the touch screen very responsible but makes it hard to click on icons and links since the chance is high that during the click the user moves the finger a few pixels and it will be registred as a swipe instead.

But we only want it to do the gesture init and not handle gestures in the handle functions.

It looks like Sony hasn't fixed this in the latest ICS update. Perhaps the consider it better to have a 9 pixel touch responsiveness and easy to click on icons.

If patch is accepted I will submit it to the CM9 branch too since it's broken there as well after the update to 4.1.B.0.431 source.

http://code.google.com/p/freexperia/issues/detail?id=766 and http://code.google.com/p/freexperia/issues/detail?id=493 describes the problem in more detail.

http://talk.sonymobile.com/thread/40474 is a complaint against the stock firmware.

Patch-set: 1