Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

We did discuss this previously and agreed not to include governors that might not be either well tested or unstable. Please note that saying the battery is better is very subjective and very hard to prove or disprove.

Even across CM maintainers there is a strong bias against including potentially unstable governors that might compromise stability in the long run and Smartass falls on that category.

Patch-set: 1